formating numbers
git-svn-id: http://emr.moh.gov.rw/svn/modules/moh-billing/branches/1.6.x@3192 6101c9c6-cf70-4d46-9f71-e9b75042c301

+++ src/org/openmrs/module/mohbilling/businesslogic/MohBillingTagUtil.java
@@ -76,7 +76,7 @@ public class MohBillingTagUtil {
 				// in order to reduce number of papers and users switching pages
 				double amountDueByPatient = 0.0;
 				for (PatientServiceBill psb : pb.getBillItems()) {
-					Double cost = psb.getUnitPrice().doubleValue()*psb.getQuantity();
+					Double cost = psb.getUnitPrice().doubleValue()*psb.getQuantity().doubleValue();
 					amountDueByPatient+=psb.getService().getFacilityServicePrice().getCategory().equals("AUTRES")?cost:cost*patientRate.doubleValue();
 				}
 //				double amountDueByPatient = (pb.getAmount().doubleValue() * patientRate
